Move code in geometry::types into geometry
