+++ src/physics/force.rs
@@ -1,5 +1,5 @@
 use super::Body;
-use crate::geometry::types::{Point, Vector};
+use crate::geometry::{Point, Vector};
 
 // Gravity ///////////////////////////////////////////////////////////////////
 //
@@ -63,7 +63,7 @@ impl Attractor {
 #[cfg(test)]
 mod tests {
     use super::{Gravity, Attractor, Body};
-    use crate::geometry::types::{Point, Vector};
+    use crate::geometry::{Point, Vector};
 
     #[test]
     #[should_panic(expected = "The minimum gravitational distance must be greater than 0.")]
