+++ src/util/gens.rs
@@ -5,8 +5,8 @@ use rand::Rng;
 use rand::thread_rng;
 use rand::ThreadRng;
 
-use crate::geometry::types::Point;
-use crate::geometry::types::Vector;
+use crate::geometry::Point;
+use crate::geometry::Vector;
 use crate::geometry::util::Transformation;
 use crate::physics::Mass;
 
@@ -227,8 +227,8 @@ impl Generator for RadialGen {
 mod tests {
     use std::f32::consts::PI;
 
-    use crate::geometry::types::Point;
-    use crate::geometry::types::Vector;
+    use crate::geometry::Point;
+    use crate::geometry::Vector;
     use crate::physics::Mass;
     use crate::util::gens::*;
